create lex.py
this implementation uses rply as its main dependency. Rply is a port of David Beasley's PLY, a python implementation of lex and yacc parsing tools. In this file, we define possible tokens that may appear in our toy-language. We also supply an action for each token name. We can also ignore whitespace in this implementation. Then we call lg.build to set up the lexer